Improve fixtures
